chore: adjust model factory to support units
